// src/generated/models/Mserp_purchpurchaseorderheaderv2entitiesModel.ts
export const Mserp_purchpurchaseorderheaderv2entitiesmserp_documentapprovalstatus = {
  200000000: 'Draft',
  200000001: 'Inreview',
  200000002: 'Rejected',
  200000003: 'Approved',
  200000004: 'Inexternalreview',
  200000005: 'Finalized',
  200000006: 'Confirmed'
} as const;
export type Mserp_purchpurchaseorderheaderv2entitiesmserp_documentapprovalstatus = keyof typeof Mserp_purchpurchaseorderheaderv2entitiesmserp_documentapprovalstatus;
// ...
export const Mserp_purchpurchaseorderheaderv2entitiesmserp_purchaseorderheadercreationmethod = {
  200000000: 'Purchase',
  200000001: 'Consignment'
} as const;
export type Mserp_purchpurchaseorderheaderv2entitiesmserp_purchaseorderheadercreationmethod = keyof typeof Mserp_purchpurchaseorderheaderv2entitiesmserp_purchaseorderheadercreationmethod;
export const Mserp_purchpurchaseorderheaderv2entitiesmserp_purchaseorderstatus = {
  200000000: 'None',
  200000001: 'Openorder',
  200000002: 'Received',
  200000003: 'Invoiced',
  200000004: 'Canceled'
} as const;
export type Mserp_purchpurchaseorderheaderv2entitiesmserp_purchaseorderstatus = keyof typeof Mserp_purchpurchaseorderheaderv2entitiesmserp_purchaseorderstatus;
// ...
export interface Mserp_purchpurchaseorderheaderv2entities extends Mserp_purchpurchaseorderheaderv2entitiesBase {
  mserp_arepricesincludingsalestaxname?: string;
  mserp_bankdocumenttypename?: string;
  mserp_dataareaid_idname?: string;
  mserp_deliveryaddresstimezonename?: string;
  mserp_documentapprovalstatus?: Mserp_purchpurchaseorderheaderv2entitiesmserp_documentapprovalstatus;
  mserp_documentapprovalstatusname?: string;
  mserp_eusaleslistcodename?: string;
  mserp_fk_invoicevendor_idname?: string;
  mserp_fk_ordervendor_idname?: string;
  mserp_formattedinvoiceaddress?: string;
  mserp_invoicetypename?: string;
  mserp_ischangemanagementactivename?: string;
  mserp_isconsolidatedinvoicetargetname?: string;
  mserp_isdelivereddirectlyname?: string;
  mserp_isdeliveryaddressorderspecificname?: string;
  mserp_isdeliveryaddressprivatename?: string;
  mserp_isonetimevendorname?: string;
  mserp_overridesalestaxname?: string;
  mserp_purchaseorderheadercreationmethod?: Mserp_purchpurchaseorderheaderv2entitiesmserp_purchaseorderheadercreationmethod;
  mserp_purchaseorderheadercreationmethodname?: string;
  mserp_purchaseorderstatus?: Mserp_purchpurchaseorderheaderv2entitiesmserp_purchaseorderstatus;
  mserp_purchaseorderstatusname?: string;
  mserp_vendortransactionsettlementtypename?: string;
  mserp_dataareaid_id?: object;
  _mserp_dataareaid_id_value?: string;
  mserp_fk_invoicevendor_id?: object;
  _mserp_fk_invoicevendor_id_value?: string;
  mserp_fk_ordervendor_id?: object;
  _mserp_fk_ordervendor_id_value?: string;
}