// src/generated/models/Mserp_purchpurchaseorderlinev2entitiesModel.ts
export const Mserp_purchpurchaseorderlinev2entitiesmserp_isaddedbychannel = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_purchpurchaseorderlinev2entitiesmserp_isaddedbychannel = keyof typeof Mserp_purchpurchaseorderlinev2entitiesmserp_isaddedbychannel;
export const Mserp_purchpurchaseorderlinev2entitiesmserp_isdeleted = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_purchpurchaseorderlinev2entitiesmserp_isdeleted = keyof typeof Mserp_purchpurchaseorderlinev2entitiesmserp_isdeleted;
// ...
export const Mserp_purchpurchaseorderlinev2entitiesmserp_purchaseorderlinecreationmethod = {
  200000000: 'Purchase',
  200000001: 'Consignment'
} as const;
export type Mserp_purchpurchaseorderlinev2entitiesmserp_purchaseorderlinecreationmethod = keyof typeof Mserp_purchpurchaseorderlinev2entitiesmserp_purchaseorderlinecreationmethod;
export const Mserp_purchpurchaseorderlinev2entitiesmserp_purchaseorderlinestatus = {
  200000000: 'None',
  200000001: 'Openorder',
  200000002: 'Received',
  200000003: 'Invoiced',
  200000004: 'Canceled'
} as const;
export type Mserp_purchpurchaseorderlinev2entitiesmserp_purchaseorderlinestatus = keyof typeof Mserp_purchpurchaseorderlinev2entitiesmserp_purchaseorderlinestatus;
// ...
export const Mserp_purchpurchaseorderlinev2entitiesmserp_workflowstate = {
  200000000: 'Notsubmitted',
  200000001: 'Submitted',
  200000002: 'Pendingapproval',
  200000003: 'Pendingcompletion',
  200000004: 'Returned',
  200000005: 'Changerequested',
  200000006: 'Completed',
  200000007: 'Approved'
} as const;
export type Mserp_purchpurchaseorderlinev2entitiesmserp_workflowstate = keyof typeof Mserp_purchpurchaseorderlinev2entitiesmserp_workflowstate;
// ...
export interface Mserp_purchpurchaseorderlinev2entities extends Mserp_purchpurchaseorderlinev2entitiesBase {
  mserp_calculatelineamountname?: string;
  mserp_catchweightunitsymbol?: string;
  mserp_dataareaid_idname?: string;
  mserp_deliveryaddresstimezonename?: string;
  mserp_diotoperationtypename?: string;
  mserp_fixedassettransactiontypename?: string;
  mserp_fk_purchaseorderheader_idname?: string;
  mserp_gsthsttaxtypename?: string;
  mserp_inventorylotid?: string;
  mserp_isaddedbychannel?: Mserp_purchpurchaseorderlinev2entitiesmserp_isaddedbychannel;
  mserp_isaddedbychannelname?: string;
  mserp_isdeleted?: Mserp_purchpurchaseorderlinev2entitiesmserp_isdeleted;
  mserp_isdeletedname?: string;
  mserp_isdeliveryaddressorderspecificname?: string;
  mserp_isdeliveryaddressprivatename?: string;
  mserp_isintrastattriangulardealname?: string;
  mserp_islinestoppedname?: string;
  mserp_isnewfixedassetname?: string;
  mserp_ispartialdeliverypreventedname?: string;
  mserp_isprojectpaywhenpaidname?: string;
  mserp_istax1099gtradeorbusinessincomename?: string;
  mserp_istax1099spropertyorservicesname?: string;
  mserp_overridesalestaxname?: string;
  mserp_purchaseorderlinecreationmethod?: Mserp_purchpurchaseorderlinev2entitiesmserp_purchaseorderlinecreationmethod;
  mserp_purchaseorderlinecreationmethodname?: string;
  mserp_purchaseorderlinestatus?: Mserp_purchpurchaseorderlinev2entitiesmserp_purchaseorderlinestatus;
  mserp_purchaseorderlinestatusname?: string;
  mserp_purchaserequisitionid?: string;
  mserp_skipcreateautochargesname?: string;
  mserp_tax1099typename?: string;
  mserp_vendorinvoicematchingpolicyname?: string;
  mserp_willproductreceivingcrossdockproductsname?: string;
  mserp_workflowstate?: Mserp_purchpurchaseorderlinev2entitiesmserp_workflowstate;
  mserp_workflowstatename?: string;
  mserp_dataareaid_id?: object;
  _mserp_dataareaid_id_value?: string;
  mserp_fk_purchaseorderheader_id?: object;
  _mserp_fk_purchaseorderheader_id_value?: string;
}